refactor: split neon operator armor rendering
